backend/requirements.txt
# Django and core dependencies
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-decouple==3.8

# WebSocket support for real-time tracking
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# Database
psycopg2-binary==2.9.10  # For PostgreSQL support (optional)

# API and serialization
drf-spectacular==0.28.0  # For API documentation
django-filter==24.3  # For filtering support (compatible with Django 5.1)

# Authentication
djangorestframework-simplejwt==5.3.1  # JWT authentication

# Google Maps integration
googlemaps==4.10.0  # Google Maps API client
aiohttp==3.10.11  # Async HTTP client for API calls
asyncio-throttle==1.0.2  # Rate limiting for async requests

# Celery for background tasks
celery==5.4.0
django-celery-beat==2.7.0  # For scheduled tasks
django-celery-results==2.5.1  # Store task results in Django

# Utilities
Pillow==10.4.0  # Image processing
python-dateutil==2.9.0
pandas==2.2.3  # Data analysis for analytics
numpy==2.2.1  # Numerical computations

# Machine Learning (for AI prediction system)
xgboost==2.1.3  # XGBoost model for client reorder predictions
scikit-learn==1.6.1  # ML utilities and preprocessing
joblib==1.4.2  # Model serialization

# Route optimization and clustering
geopy==2.4.1  # Geocoding and distance calculations
scipy==1.15.1  # Scientific computing for clustering algorithms

# Development tools
django-debug-toolbar==5.0.0
django-extensions==3.2.3
ipython==8.32.0

# Testing
pytest==8.3.4
pytest-django==4.10.0
factory-boy==3.3.1
faker==33.1.0